+++ example/codemod_analysis_required_fixtures/is_even.dart
@@ -1,0 +1,8 @@
+// Change to isEven
+var foo = (250 + 2) % 2 == 0;
+
+// Change to isOdd
+var bar = (250 + 2) % 2 == 1;
+
+// No changes, not int modulus
+var baz = 25.0 % 2 == 0;