+++ test_fixtures/functional/before/codemod_no_patches.dart
@@ -3,8 +3,8 @@ import 'dart:io';
 import 'package:codemod/codemod.dart';
 import 'package:glob/glob.dart';
 
-void main(List<String> args) {
-  exitCode = runInteractiveCodemod(
+void main(List<String> args) async {
+  exitCode = await runInteractiveCodemod(
       filePathsFromGlob(Glob('**')), NoopSuggestor(),
       args: args);
 }